updating spdlog and solving its issues

+++ NotRed/NotRed/src/NotRed/Physics/PhysicsWrappers.cpp
@@ -88,7 +88,7 @@ namespace NR
 			physx::PxActor& actor = *actors[i];
 			Entity& entity = *(Entity*)actor.userData;
 
-			NR_CORE_INFO("PhysX Actor waking up: ID: {0}, Name: {1}", entity.GetID(), entity.GetComponent<TagComponent>().Tag);
+			NR_CORE_INFO("PhysX Actor waking up: ID: {0}, Name: {1}", (uint64_t)entity.GetID(), entity.GetComponent<TagComponent>().Tag);
 		}
 	}
 
@@ -99,7 +99,7 @@ namespace NR
 			physx::PxActor& actor = *actors[i];
 			Entity& entity = *(Entity*)actor.userData;
 
-			NR_CORE_INFO("PhysX Actor going to sleep: ID: {0}, Name: {1}", entity.GetID(), entity.GetComponent<TagComponent>().Tag);
+			NR_CORE_INFO("PhysX Actor going to sleep: ID: {0}, Name: {1}", (uint64_t)entity.GetID(), entity.GetComponent<TagComponent>().Tag);
 		}
 	}
